Дорабатывает программу
