+++ js/main.js
@@ -1,2 +1,1 @@
 import './gallery.js';
-import './miniatures.js';
